requirements.txt
torch>=2.1.0
torchvision>=0.16.0
numpy>=1.24.0
scipy>=1.11.0
scikit-learn>=1.3.0
matplotlib>=3.8.0
seaborn>=0.13.0
captum>=0.7.0
shap>=0.43.0
minigrid>=2.3.0
gymnasium>=0.29.0
stable-baselines3>=2.2.0
pandas>=2.1.0
pyyaml>=6.0
tqdm>=4.66.0
Pillow>=10.0.0
skimage>=0.0
scikit-image>=0.22.0
